@@ -1,9 +1,9 @@
-ruby '2.0.0'
+ruby '2.2.2'
 
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 4.0.0.beta'
+gem 'rails' #, '~> 4.0.0.beta'
 
 #gem 'sqlite3'
 gem 'pg'
@@ -11,20 +11,20 @@ gem 'pg'
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 4.0.0.beta'
-  gem 'coffee-rails', '~> 4.0.0.beta'
+  gem 'sass-rails' #,   '~> 4.0.0.beta'
+  gem 'coffee-rails' #, '~> 4.0.0.beta'
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', platforms: :ruby
   gem 'formtastic'
-  gem 'uglifier', '>= 1.0.3'
+  gem 'uglifier' #, '>= 1.0.3'
 end
 
 gem 'jquery-rails'
 gem 'turbolinks'
-gem 'jbuilder', '~> 1.0.1'
+gem 'jbuilder' #, '~> 1.0.1'
 gem 'twitter-bootstrap-rails'
-gem 'less'
-gem 'less-rails'
+#gem 'less'
+#gem 'less-rails'
 #gem 'openehr', github: 'skoba/openehr-ruby'
 
 # To use ActiveModel has_secure_password
@@ -40,28 +40,21 @@ gem 'unicorn'
 # gem 'debugger'
 
 
-# group :development do
-#   gem 'guard-livereload'
-#   gem 'guard-unicorn'
-# # #   gem 'livereload'
-# end
-
 group :test, :development do
   gem 'rspec'
   gem 'rspec-rails'
   gem 'spring'
-  gem 'guard'#, '1.7.0'
-  gem 'guard-rspec'#, '2.6.0'
-  gem 'guard-spring' #, '0.0.2'
-  gem 'listen'
+  gem 'guard' #, '1.7.0'
+  gem 'guard-rspec' #, '2.6.0'
+#  gem 'listen', '~> 2.7'
   gem 'libnotify'
-# # for BSD guys
-#   gem 'rb-kqueue'
   gem 'simplecov'
   gem 'database_cleaner' #, github: 'bmabey/database_cleaner'
-#  gem 'rails3-generators' #mainly for factory_girl & simple_form at this point
   gem 'capybara'
-  #  gem 'forgery'
-#  gem 'livereload'
   gem 'factory_girl_rails'
+end
+
+require 'rbconfig'
+if RbConfig::CONFIG['target_os'] =~ /(?i-mx:bsd|dragonfly)/
+  gem 'rb-kqueue', '>= 0.2'
 end
