@@ -1,8 +1,5 @@
-# Load the rails application.
+# Load the Rails application.
 require File.expand_path('../application', __FILE__)
 
-# Initialize the rails application.
-MmlCare::Application.initialize!
-
-# Mime type extension for MML
-Mime::Type.register "text/xml", :mml
+# Initialize the Rails application.
+Rails.application.initialize!
