@@ -1,4 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-#MmlCare::Application.config.session_store :encrypted_cookie_store, key: '_mml-care_session'
-MmlCare::Application.config.session_store :cookie_store, key: '_mml-care_session'
+Rails.application.config.session_store :cookie_store, key: '_mml_care_session'
